requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66.0
numpy>=1.26.0
matplotlib>=3.8.0
pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0.0
pandas>=2.1.0
seaborn>=0.13.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
albumentations>=1.4.0
segmentation-models-pytorch>=0.3.3
torchmetrics>=1.0.0
timm>=0.9.0
